+++ nvim/lua/themes.lua
@@ -1,1 +1,3 @@
-vim.cmd('colorscheme iceberg') 
+-- vim.cmd('colorscheme iceberg') 
+vim.cmd [[ set background=dark ]]
+vim.cmd [[ colorscheme vscode ]]
